base history query

+++ app/src/main/java/com/radlance/fooddelivery/data/api/response/CategoryResponse.kt
@@ -3,8 +3,6 @@ package com.radlance.fooddelivery.data.api.response
 import com.google.gson.annotations.SerializedName
 
 data class CategoryResponse(
-    @SerializedName("id")
-    val id: Long,
-    @SerializedName("title")
-    val title: String
+    @SerializedName("id") val id: Int,
+    @SerializedName("title") val title: String
 )
