+++ app/src/main/java/com/radlance/fooddelivery/data/api/response/ProductResponse.kt
@@ -3,14 +3,9 @@ package com.radlance.fooddelivery.data.api.response
 import com.google.gson.annotations.SerializedName
 
 data class ProductResponse(
-    @SerializedName("id")
-    val id: Long,
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("price")
-    val price: Double,
-    @SerializedName("imageUrl")
-    val imageUrl: String,
-    @SerializedName("category")
-    val category: CategoryResponse
+    @SerializedName("id") val id: Int,
+    @SerializedName("title") val title: String,
+    @SerializedName("price") val price: Double,
+    @SerializedName("imageUrl") val imageUrl: String,
+    @SerializedName("category") val category: CategoryResponse
 )
