+++ app/src/main/java/com/radlance/fooddelivery/domain/entity/Product.kt
@@ -5,10 +5,10 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Product(
-    val id: Long,
+    val id: Int,
     val title: String,
     val price: Double,
     val imageUrl: String,
-    val categoryId: Long
+    val categoryId: Int
 ): Parcelable
